Using chunk.size(4)

+++ 1.0-spring-batch/src/main/java/com/guru2batch/writer/FirstItemWriter.java
@@ -6,10 +6,10 @@ import org.springframework.stereotype.Component;
 import java.util.List;
 
 @Component
-public class FirstItemWriter implements ItemWriter<Integer> {
+public class FirstItemWriter implements ItemWriter<Long> {
 
     @Override
-    public void write(List<? extends Integer> items) throws Exception {
+    public void write(List<? extends Long> items) throws Exception {
         System.out.println("Inside Item Writer");
         items.stream().forEach(System.out::println);
     }
